fixed warnings
git-svn-id: https://svn.intern.artcom.de/pro60/trunk/src@3117 f3027e6b-e3f5-0310-a8e1-edde99107618

+++ Y60/scene/CameraBuilder.cpp
@@ -37,7 +37,7 @@ namespace y60 {
     }
 
     void
-    CameraBuilder::setHFov(const double theViewingAngle) {
+    CameraBuilder::setHFov(const float theViewingAngle) {
         if (getNode()->hasFacade()) {
             getNode()->getFacade<Camera>()->set<HfovTag>(theViewingAngle);
         } else {
